backend/pyproject.toml
[tool.poetry]
name = "ai_summarizer"
version = "0.0.1"
description = ""
authors = ["Dmitry Buslov <buslovdmitrij0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pypdf2 = "^3.0.1"
langchain-openai = "^0.1.23"
drf-yasg = "^1.21.7"
langchain-text-splitters = "^0.2.4"
strenum = "^0.4.15"
